organize dep packages in a dep subfolder
Moves deptoken and depsync into dep.
