+++ cmd/micromdm/serve.go
@@ -48,8 +48,8 @@ import (
 	"github.com/micromdm/micromdm/core/list"
 	"github.com/micromdm/micromdm/core/remove"
 	"github.com/micromdm/micromdm/crypto"
-	"github.com/micromdm/micromdm/depsync"
-	"github.com/micromdm/micromdm/deptoken"
+	"github.com/micromdm/micromdm/dep/depsync"
+	"github.com/micromdm/micromdm/dep/deptoken"
 	"github.com/micromdm/micromdm/device"
 	"github.com/micromdm/micromdm/mdm/checkin"
 	"github.com/micromdm/micromdm/mdm/connect"
