+++ core/apply/service.go
@@ -18,7 +18,7 @@ import (
 	"github.com/micromdm/dep"
 	"github.com/micromdm/micromdm/appstore"
 	"github.com/micromdm/micromdm/blueprint"
-	"github.com/micromdm/micromdm/deptoken"
+	"github.com/micromdm/micromdm/dep/deptoken"
 	"github.com/micromdm/micromdm/profile"
 	"github.com/micromdm/micromdm/pubsub"
 	"github.com/micromdm/micromdm/user"
