+++ device/db.go
@@ -6,12 +6,13 @@ import (
 	"time"
 
 	"github.com/boltdb/bolt"
-	"github.com/micromdm/micromdm/depsync"
+	"github.com/pkg/errors"
+	uuid "github.com/satori/go.uuid"
+
+	"github.com/micromdm/micromdm/dep/depsync"
 	"github.com/micromdm/micromdm/mdm/checkin"
 	"github.com/micromdm/micromdm/mdm/connect"
 	"github.com/micromdm/micromdm/pubsub"
-	"github.com/pkg/errors"
-	uuid "github.com/satori/go.uuid"
 )
 
 const (
